Created a new class RiotRamWidget, which takes the place of the old RamWidget.  Class RawWidget is now an abstract base class to RiotRamWidget and CartRamWidget, allowing to share quite a lot of code.  In the process, both classes now implement input from the binary and decimal widgets, not just the DataGridWidget.
git-svn-id: svn://svn.code.sf.net/p/stella/code/trunk@3195 8b62c5a3-ac7e-4cc8-8f21-d9a121418aba

+++ src/debugger/CartDebug.hxx
@@ -40,7 +40,7 @@ using CartMethod = int (CartDebug::*)();
 class CartState : public DebuggerState
 {
   public:
-    IntArray ram;    // The actual data values
+    ByteArray ram;   // The actual data values
     IntArray rport;  // Address for reading from RAM
     IntArray wport;  // Address for writing to RAM
     string bank;     // Current banking layout
